Support alt text for images
Generate alt text via OpenAI, and add it to the UI.

I don't love automatically generating alt text, but I also recognize that it's not realistic to write alt text for hundreds of photos. It _is_ realistic to edit that alt text when it could be better. So, in the name of accessibility, I generated a bunch of alt text.